+++ src/index.ts
@@ -5,6 +5,7 @@ import { cloudfunctionsv2Policies } from './cloudfunctionsv2'
 import { cloudrunPolicies } from './cloudrun'
 import { cloudrunv2Policies } from './cloudrunv2'
 import { computePolicies } from './compute'
+import { containerPolicies } from './container'
 import { sqlPolicies } from './sql'
 import { storagePolicies } from './storage'
 
@@ -15,6 +16,7 @@ const allPolicies = [
 	...cloudrunPolicies,
 	...cloudrunv2Policies,
 	...computePolicies,
+	...containerPolicies,
 	...sqlPolicies,
 	...storagePolicies
 ]